change(components/Modal): Split into Modal and Dialog
