+++ src/components/Modal/Modal.js
@@ -1,84 +1,8 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import styled, { css } from 'styled-components';
+import styled from 'styled-components';
 import BaseModal from 'react-modal2';
 import { DESKTOP } from '../../constants/Breakpoints';
-import { hexToRgba } from '../../helpers/colors';
-
-const PADDINGS = {
-  small: '16px 24px',
-  normal: '24px 48px',
-  large: '24px 48px',
-};
-
-const WIDTHS = {
-  small: 'width: 300px',
-  normal: 'width: 600px',
-  large: 'width: 900px',
-  auto: `
-    width: auto;
-    max-width: 900px;
-  `,
-};
-
-const HeaderStyled = styled.header.attrs({ tabIndex: 0 })`
-  outline: none;
-`;
-
-const FooterStyled = styled.footer``;
-
-const BodyStyled = styled.div``;
-
-export const ModalStyles = css`
-  background-color: ${props => props.theme.white};
-  box-shadow: 0 0 8px 0 rgba(120, 130, 139, 0.5);
-  width: 100%;
-  max-width: 100%;
-  @media ${DESKTOP} {
-    ${props => WIDTHS[props.size]};
-  }
-
-  ${HeaderStyled} {
-    border-bottom: solid 1px ${props => hexToRgba(props.theme.stoneGrey, 0.5)};
-    padding: ${props => PADDINGS[props.size] || PADDINGS.normal};
-    font-size: 1.25em;
-    font-weight: bold;
-    text-align: center;
-    color: ${props => props.theme.skyBlue};
-    @media ${DESKTOP} {
-      font-size: ${props => (props.size === 'small' ? '1.25em' : '1.5em')};
-    }
-  }
-
-  ${FooterStyled} {
-    border-top: solid 1px ${props => hexToRgba(props.theme.stoneGrey, 0.5)};
-    padding: ${props => PADDINGS[props.size] || PADDINGS.normal};
-    display: flex;
-    flex-flow: column nowrap;
-    align-items: center;
-
-    > * + * {
-      margin-top: 20px;
-    }
-
-    ${props =>
-    props.size !== 'small'
-      && `
-      @media ${DESKTOP} {
-        flex-flow: row-reverse wrap;
-        align-items: center;
-        justify-content: space-between;
-        > * + * {
-          margin-top: 0;
-        }
-      }
-    `};
-  }
-
-  ${BodyStyled} {
-    padding: ${props => PADDINGS[props.size] || PADDINGS.normal};
-  }
-`;
 
 const Root = styled.section`
   .ft-modal--backdrop {
@@ -99,19 +23,20 @@ const Root = styled.section`
     z-index: 10000;
     position: absolute;
     max-height: 100vh;
+    width: 100%;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
     &:focus {
       outline: none;
     }
     @media ${DESKTOP} {
-      top: 50%;
-      left: 50%;
-      transform: translate(-50%, -50%);
+      width: auto;
     }
   }
 `;
 
-const ModalWrapper = styled.div`
-  ${ModalStyles};
+const Content = styled.div`
   height: 100vh;
   @media ${DESKTOP} {
     height: auto;
@@ -120,7 +45,7 @@ const ModalWrapper = styled.div`
 `;
 
 /**
- * Main modal component.
+ * Modal component.
  */
 class Modal extends Component {
   preventBodyOverflow = null;
@@ -157,12 +82,12 @@ class Modal extends Component {
   }
 
   render() {
-    const { backdropColor, open, size, children, onClose } = this.props;
+    const { backdropColor, open, children, onClose } = this.props;
     if (!open) {
       return null;
     }
     return (
-      <Root backdropColor={backdropColor} size={size} role="dialog">
+      <Root backdropColor={backdropColor} role="dialog">
         <BaseModal
           onClose={onClose}
           closeOnEsc
@@ -170,7 +95,7 @@ class Modal extends Component {
           backdropClassName="ft-modal--backdrop"
           modalClassName="ft--modal--dialog"
         >
-          <ModalWrapper size={size}>{children}</ModalWrapper>
+          <Content>{children}</Content>
         </BaseModal>
       </Root>
     );
@@ -179,7 +104,6 @@ class Modal extends Component {
 
 Modal.defaultProps = {
   preventGlobalScroll: true,
-  size: 'normal',
 };
 
 Modal.propTypes = {
@@ -187,13 +111,8 @@ Modal.propTypes = {
   children: PropTypes.node.isRequired,
   preventGlobalScroll: PropTypes.bool,
   open: PropTypes.bool.isRequired,
-  size: PropTypes.oneOf(['auto', 'small', 'normal', 'large']),
   onClose: PropTypes.func,
 };
-
-Modal.Header = HeaderStyled;
-Modal.Footer = FooterStyled;
-Modal.Body = BodyStyled;
 
 /** @component */
 export default Modal;
