+++ src/components/index.js
@@ -1,11 +1,12 @@
 export Badge from './Badge';
 export Button from './Button';
-export ColorSelector from './ColorSelector';
-export Dropdown from './Dropdown';
-export OverflowMenu from './OverflowMenu';
-export Icon from './Icon';
-export Table from './Table';
-export LoadingSpinner from './LoadingSpinner';
-export Searchbar from './Searchbar';
-export Modal from './Modal';
 export CheckBox from './CheckBox';
+export ColorSelector from './ColorSelector';
+export Dialog from './Dialog';
+export Dropdown from './Dropdown';
+export Icon from './Icon';
+export LoadingSpinner from './LoadingSpinner';
+export Modal from './Modal';
+export OverflowMenu from './OverflowMenu';
+export Searchbar from './Searchbar';
+export Table from './Table';
